Add Script Extender launchers
modded .exes that launch the script extender exe for their respective games. all still compatible with the launcher editor. each one required a different version of a similar hack, and for fallout 3 i started with the epic games launcher, before realising it's kind of terrible when compared to the steam one, and i had to start from scratch there. whatever, it's done now.